[SASP-3014] JarJar should not mangle non-modified jars (#1)
Modifies jarjar so that it tracks whether the content is modified and if not, the class file is returned unmodified.

Purpose of this is making it easier to deal with class path conflict errors. Before this change, if the shading is applied to certain bazel target, then even if some dependency is not subject to shading rules, it gets mangled in non-compatible way with original library, upsetting classpath checker, when there is no actual problem.

Originally developed in in https://github.com/databricks/universe/pull/464965/commits/1f4b9cc17395746af7beb7313ed0044a5dbd5bc6

Test coverage: unit tests.