Escape $ in patterns
`$` is a valid character in an identifier (`Character.isJavaIdentifierPart`) and thus it is (correctly) accepted by jarjar in rule patterns, but it is not escaped when building the regex (which is not necessary for other identifier characters). This makes it impossible to match inner classes (which always contain a `$`) and other synthetic classes directly (which is necessary to rename classes instead of simply moving them to other packages). Escaping it as `\$` makes these matches work correctly.